instances crimini_mushrooms and sliced_shiitake_mushrooms were NOT merged together
Query: in the Edible Mushrooms class, should the instances 'Crimini mushrooms' and 'Sliced shiitake mushrooms' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

Crimini mushrooms and sliced shiitake mushrooms are two different types of mushrooms with distinct characteristics, flavors, and textures. While they share some similarities, they are not interchangeable and can be used in different recipes and dishes. Therefore, it is not appropriate to merge them as semantic and ontological duplicates.
meaning that instances crimini_mushrooms and sliced_shiitake_mushrooms are different.
Query cache in files:
- /kgfiller/data/cache-59171581a375d9c40dbea39ba490721a5e54d20df59e1292731b006b336b5eea.yml