instances tomato and diced_tomatoes were NOT merged together
Query: in the Edible Vegetable class, should the instances 'Tomato' and 'diced tomatoes' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

The reason is that "tomato" refers to the whole fruit, while "diced tomatoes" refers to the fruit that has been processed into small pieces. While they are related, they have different meanings and uses in cooking and nutrition. Therefore, it is not appropriate to merge them as semantic and ontological duplicates.
meaning that instances tomato and diced_tomatoes are different.
Query cache in files:
- /kgfiller/data/cache-75f15d97571c854c337dca6044e25108e0484b6aab30ca2ae7ebaaecf7747613.yml