instances diced_tomatoes and zucchini__sliced were NOT merged together
Query: in the Edible Vegetable class, should the instances 'Zucchini, sliced' and 'diced tomatoes' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

The terms "sliced zucchini" and "diced tomatoes" describe different forms of vegetables, even though they are both edible and belong to the same category of vegetables. Therefore, they should not be merged as semantic and ontological duplicates.
meaning that instances diced_tomatoes and zucchini__sliced are different.
Query cache in files:
- /kgfiller/data/cache-448212bb683f5eb04b0eadc9abde54bd9938892b396418f1067a7dfb94ce0bf5.yml