instances lime_juice and freshly_squeezed_key_lime_juice were NOT merged together
Query: in the Juice class, should the instances 'Freshly squeezed key lime juice' and 'Lime juice' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

The two instances are not identical in their description, with "Freshly squeezed" providing additional information about the preparation method of the juice. This difference in detail makes them distinct entities, and they should not be merged as semantic and ontological duplicates.
meaning that instances lime_juice and freshly_squeezed_key_lime_juice are different.
Query cache in files:
- /kgfiller/data/cache-fc317fb9ae3d2d484e3e0ccb21bb16d5401bab7b82063347a3fb9ffadd13c2ab.yml