instances __the_base_ingredient_of_the_sauce and tomato_sauce were NOT merged together
Query: in the Sauce class, should the instances ': The base ingredient of the sauce' and 'Tomato sauce' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

The instance ": The base ingredient of the sauce" refers to the main component used in creating any type of sauce, while "Tomato sauce" is a specific type of sauce made using tomatoes as its primary ingredient. Therefore, they are not semantic nor ontological duplicates.
meaning that instances __the_base_ingredient_of_the_sauce and tomato_sauce are different.
Query cache in files:
- /kgfiller/data/cache-2f38c655fd65606b0b7cf59b81c54c9f8e8e7d3490cf1b0acf5939e82bada89e.yml